Create methods to view data representation
Create functions which print binary and hexadecimal forms of a given number.

There is an adjustment for endianness (the small endian is assumed), but other than that the functions do not alter the number in any way (e.g., a negative number will be printed out in the two's complement form).